Move CacheConfig under config package
