+++ src/main/java/uk/gov/hmcts/opal/config/CacheConfig.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.cache;
+package uk.gov.hmcts.opal.config;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
